now built as static library

+++ Ares/src/Ares/Events/MouseEvent.h
@@ -10,8 +10,8 @@ namespace Ares {
 		MouseMovedEvent(float x, float y)
 			: m_MouseX(x), m_MouseY(y) { }
 
-		inline int GetX() const { return m_MouseX; }
-		inline int GetY() const { return m_MouseY; }
+		inline float GetX() const { return m_MouseX; }
+		inline float GetY() const { return m_MouseY; }
 
 		std::string ToString() const override
 		{
@@ -33,8 +33,8 @@ namespace Ares {
 		MouseScrolledEvent(float xOffset, float yOffset)
 			: m_XOffset(xOffset), m_YOffset(yOffset) { }
 
-		inline int GetXOffset() const { return m_XOffset; }
-		inline int GetYOffset() const { return m_YOffset; }
+		inline float GetXOffset() const { return m_XOffset; }
+		inline float GetYOffset() const { return m_YOffset; }
 
 		std::string ToString() const override
 		{
